% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pride_palette.R
\name{pride_palette}
\alias{pride_palette}
\title{Pride Flag Color Schemes}
\usage{
pride_palette(palette)
}
\arguments{
\item{palette}{[\code{character(1)}]\cr
Which Pride flag would you like to use as the basis of your color scheme?}
}
\description{
This function contains a set of palettes based on the colors of various Pride
flags. The function's only argument (\code{palette}) is used to select the flag
you want to use as the basis of your color scheme. This function works well
wrapped inside the \code{scale_color_manual} and \code{scale_fill_manual}
functions in \pkg{ggplot2}.
}
\details{
Available color schemes are:
\itemize{
\item \code{"pride"} -- the six-color Pride Flag
\item \code{"philly_poc_pride"} -- the Philadelphia People Of Color Pride Flag
\item \code{"gilbert_baker_pride"} -- Gilbert Baker's 1978 Pride Flag
\item \code{"bisexual_pride"} -- the Bisexual Pride Flag
\item \code{"pansexual_pride"} -- the Pansexual Pride Flag
\item \code{"asexual_pride"} -- the Asexual Pride Flag
\item \code{"trans_pride"} -- the Trans Pride Flag
\item \code{"genderfluid_pride"} -- the Genderfluid Pride Flag
\item \code{"genderqueer_pride"} -- the Genderqueer Pride Flag
\item \code{"polysexual_pride"} -- the Polysexual Pride Flag
\item \code{"agender_pride"} -- the Agender Pride Flag
\item \code{"aromantic_pride"} -- the Aromantic Pride Flag
\item \code{"nonbinary_pride"} -- the Nonbinary Pride Flag
}
}
\examples{
# A scatterplot using the Philadelphia People Of Color Pride Flag

var1 <- rnorm(n = 80)
var2 <- rnorm(n = 80)
var3 <- as.factor(rep(1:8, times = 10))
data <- data.frame(var1, var2, var3)

ggplot(data = data, mapping = aes(x = var1, y = var2)) +
  scale_color_manual(values = pride_palette(palette = "philly_poc_pride")) +
  geom_point(aes(color = var3)) +
  theme_minimal()

}
\author{
Joel Le Forestier (Twitter: @JoelLeForestier; Website: joelleforestier.com; Email: joel.leforestier@mail.utoronto.ca)
}
